requirements.txt
torch
torchvision
transformers
pycocotools
opencv-python
pillow
numpy
pandas
matplotlib
pyyaml
tqdm
jupyter
ipykernel
EOF
